User Story 1 e 2 Completate

+++ database/migrations/2023_11_21_201407_create_products_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('title');
             $table->text('description');
             $table->string('cover')->default('default/product-auto-img.png');
-            $table->float('price', 5, 2);
+            $table->decimal('price', 8, 2);
             $table->timestamps();
         });
     }
